pyproject.toml
[project]
name = "artem-shumeiko-pet-project-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "celery>=5.5.3",
    "fastapi-cache2>=0.2.2",
    "notebook>=7.4.3",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pydantic2>=1.1.15",
    "pyjwt>=2.10.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.42",
]
